fix: move mocks to /src/test
