+++ nest-cli.json
@@ -3,6 +3,6 @@
   "collection": "@nestjs/schematics",
   "sourceRoot": "src",
   "compilerOptions": {
-    "assets": [{ "include": "i18n/**/*", "watchAssets": true, "outDir": "dist/src/i18n" }]
+    "assets": [{ "include": "i18n/**/*", "watchAssets": true }]
   }
 }
