+++ package.json
@@ -13,8 +13,8 @@
     "migration:create": "npm run typeorm -- migration:create",
     "migration:run": "npm run typeorm -- --dataSource=src/database/data-source.ts migration:run",
     "migration:revert": "npm run typeorm -- --dataSource=src/database/data-source.ts migration:revert",
-    "migration:run:prod": "ts-node -r tsconfig-paths/register  ./dist/src/database/run-multi-tenant-migrations.js",
-    "seed:run:prod": "ts-node -r tsconfig-paths/register ./dist/src/database/seeds/relational/run-seed.js",
+    "migration:run:prod": "ts-node -r tsconfig-paths/register  ./dist/database/run-multi-tenant-migrations.js",
+    "seed:run:prod": "ts-node -r tsconfig-paths/register ./dist/database/seeds/relational/run-seed.js",
     "schema:drop": "npm run typeorm -- --dataSource=src/database/data-source.ts schema:drop",
     "schema:reset": "env-cmd ts-node -r tsconfig-paths/register ./src/database/run-schema-reset.ts",
     "seed:create:relational": "hygen seeds create-relational",
